virt-operator: share KubeVirt semantic validation with reconciler
Extract the pure validators into a shared package and call them from Sync so invalid specs fail the rollout with a Degraded condition and event.
