+++ pkg/virt-operator/resource/apply/BUILD.bazel
@@ -37,6 +37,7 @@ go_library(
         "//pkg/virt-operator/resource/generate/rbac:go_default_library",
         "//pkg/virt-operator/resource/placement:go_default_library",
         "//pkg/virt-operator/util:go_default_library",
+        "//pkg/virt-operator/validation:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/api/instancetype/v1beta1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
