+++ pkg/virt-operator/validation/validation_suite_test.go
@@ -1,0 +1,11 @@
+package validation_test
+
+import (
+	"testing"
+
+	"kubevirt.io/client-go/testutils"
+)
+
+func TestValidation(t *testing.T) {
+	testutils.KubeVirtTestSuiteSetup(t)
+}
